Move add_minimize_ellipsoid_volume to ellipsoid_utils.py
